Fully implement starting reminders on boot
As it turned out, before I think it only created the notifictaion at
boot and nothing else?? Plus, the boot receiver was _always_ registered,
even if starting on boot was disabled, which is kind of wasteful. With
this, the receiver is only enabled when starting on boot is actually
enabled, and it actually works too.